requirements.txt
# Direct dependencies
python-dotenv>=1.0.0
openai>=2.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.0.0
